+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.5.15
 
+- use > instead of >= in SBML and Simbio events to support run at start
+- minor bug fixess
 
 ## 0.5.14